Function formdata::parse_multipart

pub fn parse_multipart(request: &mut Request) -> Result<(Vec<(String, String)>, Vec<(String, UploadedFile)>), Error>

This function parses and processes hyper::server::Request data that contains multipart/form-data content. It does this in a streaming fashion, saving embedded uploaded files to disk as they are encountered by the parser.

It returns two sets of data. The first, Vec<(String,String)>, are the variable-value pairs from the POST-ed form. The second Vec<(String,UploadedFile)> are the variable-file pairs from the POST-ed form, where UploadedFile structures describe the uploaded file.