Struct formdata::FormData

pub struct FormData {
    pub fields: Vec<(String, String)>,
    pub files: Vec<(String, FilePart)>,
}

The extracted text fields and uploaded files from a multipart/form-data request.

Use parse_multipart to devise this object from a request.

Fields
fields: Vec<(String, String)>

Name-value pairs for plain text fields. Technically, these are form data parts with no filename specified in the part's Content-Disposition.

files: Vec<(String, FilePart)>

Name-value pairs for temporary files. Technically, these are form data parts with a filename specified in the part's Content-Disposition.

Methods

impl FormData

fn new() -> FormData

fn to_multipart(&self) -> Result<Vec<Node>, Error>

Create a mime-multipart Vec from this FormData