Struct formdata::UploadedFile

pub struct UploadedFile {
    pub path: PathBuf,
    pub filename: Option<String>,
    pub content_type: Mime,
    pub size: usize,
}

An uploaded file that was received as part of multipart/form-data parsing.

Files are streamed to disk because they may not fit in memory.

Fields

path

The temporary file where the data was saved.

filename

The filename that was specified in the data, unfiltered. It may or may not be legal on the local filesystem.

content_type

The unvalidated content-type that was specified in the data.

size

The size of the file.