Crate formdata

This crate parses and processes hyper::server::Request data that contains multipart/form-data content.

The main entry point is parse_multipart

Reexports

pub use error::Error;

Modules

buf

error

Structs

FormData
The extracted text fields and uploaded files from a multipart/form-data request.

UploadedFile
An uploaded file that was received as part of multipart/form-data parsing.

Traits

Request
A wrapper for request data to provide parsing multipart requests to any front-end that provides a hyper::header::Headers and a std::io::Read of the request's entire body.

Functions

parse_multipart
Parses and processes hyper::server::Request data that is multipart/form-data content.