Crate formdata

This crate parses and processes hyper::server::Request data that contains multipart/form-data content.

The main entry point is parse_multipart

Reexports

pub use error::Error;

pub use buf::BufReadPlus;

Modules

buf

error

Structs

ContentDispositionFormData
We define a Content-Disposition form-data only header, which is found within the mime multipart sections. NOT FOR GENERAL USE

UploadedFile
This structure represents uploaded files which were received as part of the multipart/form-data parsing. They are streamed to disk because they may not fit in memory.

Traits

Request
A wrapper trait for hyper::server::Request data to provide parsing multipart requests to any front-end that provides a hyper::header::Headers and a std::io::Read of the request's entire body.

Functions

parse_multipart
This function parses and processes hyper::server::Request data that contains multipart/form-data content. It does this in a streaming fashion, saving embedded uploaded files to disk as they are encountered by the parser.