Enum formdata::Error

pub enum Error {
    NoRequestContentType,
    NotMultipart,
    NotFormData,
    BoundaryNotSpecified,
    PartialHeaders,
    MissingDisposition,
    InvalidDisposition,
    NoName,
    Eof,
    Httparse(Error),
    Io(Error),
    Hyper(Error),
    Utf8(FromUtf8Error),
    Decoding(Cow<'static, str>),
    Multipart(Error),
}

An error type for the formdata crate.

Variants
NoRequestContentType

The Hyper request did not have a Content-Type header.

NotMultipart

The Hyper request Content-Type top-level Mime was not Multipart.

NotFormData

The Hyper request Content-Type sub-level Mime was not FormData.

BoundaryNotSpecified

The Content-Type header failed to specify boundary token.

PartialHeaders

A multipart section contained only partial headers.

MissingDisposition

A multipart section did not have the required Content-Disposition header.

InvalidDisposition

A multipart section did not have a valid corresponding Content-Disposition.

NoName

A multipart section Content-Disposition header failed to specify a name.

Eof

The request body ended prior to reaching the expected terminating boundary.

Httparse(Error)

An HTTP parsing error from a multipart section.

Io(Error)

An I/O error.

Hyper(Error)

An error was returned from Hyper.

Utf8(FromUtf8Error)

An error occurred during UTF-8 processing.

Decoding(Cow<'static, str>)

An error occurred during character decoding

Multipart(Error)

A MIME multipart error